+++ extensions/debug/views/default/panels/request/summary.php
@@ -12,7 +12,7 @@ if ($statusCode === null) {
 }
 if ($statusCode >= 200 && $statusCode < 300) {
     $class = 'label-success';
-} elseif ($statusCode >= 100 && $statusCode < 200) {
+} elseif ($statusCode >= 300 && $statusCode < 400) {
     $class = 'label-info';
 } else {
     $class = 'label-important';