+++ extensions/gii/generators/model/default/model.php
@@ -41,6 +41,16 @@ class <?= $className ?> extends <?= '\\' . ltrim($generator->baseClass, '\\') . 
     {
         return '<?= $tableName ?>';
     }
+<?php if ($generator->db !== 'db'): ?>
+
+    /**
+     * @return \yii\db\Connection the database connection used by this AR class.
+     */
+    public static function getDb()
+    {
+        return Yii::$app->get('<?= $generator->db ?>');
+    }
+<?php endif; ?>
 
     /**
      * @inheritdoc