+++ framework/db/Connection.php
@@ -528,14 +528,17 @@ class Connection extends Component
      */
     public function quoteSql($sql)
     {
-        return preg_replace_callback('/(\\{\\{(%?[\w\-\. ]+%?)\\}\\}|\\[\\[([\w\-\. ]+)\\]\\])/',
+        return preg_replace_callback(
+            '/(\\{\\{(%?[\w\-\. ]+%?)\\}\\}|\\[\\[([\w\-\. ]+)\\]\\])/',
             function ($matches) {
                 if (isset($matches[3])) {
                     return $this->quoteColumnName($matches[3]);
                 } else {
                     return str_replace('%', $this->tablePrefix, $this->quoteTableName($matches[2]));
                 }
-            }, $sql);
+            },
+            $sql
+        );
     }
 
     /**