+++ framework/rbac/Rule.php
@@ -33,9 +33,11 @@ abstract class Rule extends Object
     /**
      * Executes the rule.
      *
-     * @param Item $item the auth item that this rule is associated with
-     * @param array $params parameters passed to [[ManagerInterface::allow()]].
+     * @param string|integer $user the user ID. This should be either an integer or a string representing
+     * the unique identifier of a user. See [[\yii\web\User::id]].
+     * @param Item $item the role or permission that this rule is associated with
+     * @param array $params parameters passed to [[ManagerInterface::checkAccess()]].
      * @return boolean a value indicating whether the rule permits the auth item it is associated with.
      */
-    abstract public function execute($item, $params);
+    abstract public function execute($user, $item, $params);
 }